Create PE file analyzer
This is a script that analyzes a PE file (a Windows executable file) by extracting its functions, generating control flow graphs, and performing symbolic execution on the opcodes in the longest path of the control flow graph, or a random path if the number
of edges in the graph is greater than 100. It also checks if the call target of any opcode is an API call, and if so, looks up the import table to get the name of the function and number of arguments, and puts this information in a database. It also updates the EAX register. It also accepts options for --help and --verbose.